+++ src/kio/schema/metadata/v12/response.py
@@ -79,7 +79,7 @@ class MetadataResponse:
     cluster_id: str | None = field(metadata={"kafka_type": "string"}, default=None)
     """The cluster ID that responding broker belongs to."""
     controller_id: BrokerId = field(
-        metadata={"kafka_type": "int32"}, default=BrokerId(i32(-1))
+        metadata={"kafka_type": "int32"}, default=BrokerId(-1)
     )
     """The ID of the controller broker."""
     topics: tuple[MetadataResponseTopic, ...]
